Broke singleton out as a separate word.
Also added inheritance of answers for defining words.

And deleted duplicate code in ucc/word/word.py setting the filename_suffix
attribute on defining words (it now belongs in top_package where it can be
done after the kind_obj attributes are set, which are needed for the answer
inheritance).